Move `get_system_locale()` to `locale` module
