+++ crates/nu-utils/src/lib.rs
@@ -1,6 +1,8 @@
+pub mod locale;
 pub mod utils;
 
+pub use locale::get_system_locale;
 pub use utils::{
-    enable_vt_processing, get_default_config, get_default_env, get_system_locale, get_ls_colors,
+    enable_vt_processing, get_default_config, get_default_env, get_ls_colors,
     stderr_write_all_and_flush, stdout_write_all_and_flush,
 };
